// src/app/Model/AuthTable.ts
   export interface AuthTable {
  get_status_count: Getstatuscount;
  registered: number;
  name: string;
  hasPhoto: string;
  phone_number: string;
  id: string;
  is_lead: number;
  status: string;
  apiwha_autoreply: string;
  buttons: string;
  show_ad: number;
  paid_status: string;
  partial_data: Partialdata;
  photo?: any;
  is_premium: string;
  today_profile_count: string;
  user_income: number;
  family_income: number;
  birth_time: string;
  birth_place: string;
  birth_date: string;
  gender: string;
}

   export interface Partialdata {
  id: number;
  is_approve_ready: number;
  created_at: string;
  updated_at: string;
  mobile: string;
  new_mobile: string;
  temple_id: string;
  age: string;
  photo?: any;
  photo_score?: any;
  carousel?: any;
  religion: string;
  caste: string;
  education: string;
  height: string;
  weight: number;
  manglik: string;
  marital_status: string;
  occupation: string;
  gender: string;
  annual_income: number;
  birth_date: string;
  birth_time: string;
  locality: string;
  degree: string;
  college: string;
  additional_qualification: string;
  profession: string;
  company: string;
  working_city: string;
  food_choice: string;
  about: string;
  lead_id: string;
  extra_lead_id: string;
  backup_lead_id: string;
  org_lead_id: string;
  profiles_sent: number;
  sent_profiles?: any;
  pdf_url?: any;
  wantBot: number;
  bot_language: string;
  fcm_id: string;
  bot_status: string;
  last_seen: string;
  fcm_app: string;
  birth_place: string;
  relation: string;
  lat_locality?: any;
  long_locality?: any;
  education_score: string;
  is_premium_interest?: any;
  profile_percent: number;
  disability?: any;
  disabled_part?: any;
  abroad?: any;
  si_event: number;
  unapprove_carousel?: any;
  is_call_back: number;
  call_back_query?: any;
  is_deleted: number;
  deleted_by?: any;
  name: string;
  online_score: number;
  family: FamilyAuth;
  preferences: PreferencesAtuh;
}

   export interface PreferencesAtuh {
  id: number;
  lead_id: number;
  age_min: number;
  age_max: number;
  height_min: string;
  height_max: string;
  caste: string;
  marital_status: string;
  manglik: string;
  food_choice: string;
  working: string;
  occupation: string;
  income_min: number;
  income_max: number;
  citizenship?: any;
  description: string;
  caste_no_bar?: any;
  mother_tongue: string;
  created_at: string;
  updated_at: string;
  amount_collected: number;
  religion: string;
  pref_state: string;
  pref_country: string;
  pref_city: string;
  pref_state_id: string;
  pref_country_id: string;
}

   export interface FamilyAuth {
  id: number;
  lead_id: number;
  name?: any;
  relation: string;
  livingWithParents: string;
  locality: string;
  mobile: string;
  new_mobile: string;
  backup_mobile: string;
  email: string;
  unmarried_sons: number;
  married_sons: number;
  unmarried_daughters: number;
  married_daughters: number;
  family_type: string;
  house_type: string;
  religion: string;
  caste: string;
  gotra: string;
  occupation: string;
  family_income: number;
  office_address?: any;
  father_status: string;
  mother_status: string;
  whats_app_no: string;
  mother_tongue?: any;
  occupation_mother: string;
  address?: any;
  about: string;
  city: string;
  father_off_addr?: any;
  created_at: string;
  updated_at: string;
  email_verified: number;
}

   export interface Getstatuscount {
  S: number;
  SL: number;
  R: number;
  C: number;
  M: number;
}